requirements.txt
addict
loguru
pyyaml
redis
pillow
numpy
scipy
pandas
lapx